+++ example/main.dart
@@ -1,26 +1,23 @@
-# angular_sentry
+import 'dart:html' hide Event;
 
-Helper to implements sentry with Angular.
-
-## Usage
-
-### Basic
-
-```dart
 import "package:angular/angular.dart";
 import "package:angular_sentry/angular_sentry.dart";
+import 'package:http/http.dart';
 
 // ignore: uri_has_not_been_generated
 import 'main.template.dart' as ng;
 
+// ignore: uri_has_not_been_generated
+import 'app.template.dart' as app;
+
 const sentryModule = Module(provide: [
-    ValueProvider.forToken(sentryLoggerToken, "MY_SENTRY_DSN"),
-    ValueProvider.forToken(sentryEnvironmentToken, "production"),
-    ValueProvider.forToken(sentryReleaseVersionToken, "1.0.0"),
-    ClassProvider<ExceptionHandler>(
-        ExceptionHandler,
-        useClass: AngularSentry,
-      ),
+  //ValueProvider.forToken(sentryLoggerToken, "MY_SENTRY_DSN"),
+  ValueProvider.forToken(sentryEnvironmentToken, "production"),
+  ValueProvider.forToken(sentryReleaseVersionToken, "1.0.0"),
+  ClassProvider<ExceptionHandler>(
+    ExceptionHandler,
+    useClass: AngularSentry,
+  ),
 ]);
 
 @GenerateInjector(
@@ -29,34 +26,14 @@ const sentryModule = Module(provide: [
 const scannerApp = ng.scannerApp$Injector;
 
 main() {
-  runApp(appComponentNgFactory, createInjector: scannerApp);
-}
-```
-
-### Advanced
-
-Implement your own class using AngularSentry
-
-```dart
-
-const sentryModule = Module(provide: [
-    ...
-    ClassProvider<ExceptionHandler>(
-        ExceptionHandler,
-        useClass: AppSentry,
-      ),
-]);
-
-main() {
-  runApp(appComponentNgFactory, createInjector: scannerApp);
+  runApp(app.AppComponentNgFactory, createInjector: scannerApp);
 }
 
 class AppSentry extends AngularSentry {
-  AppSentry(Injector injector, NgZone zone)
+  AppSentry(Injector injector)
       : super(
           injector,
-          zone,
-          dsn: "MY_SENTRY_DSN",
+          //dsn: "MY_SENTRY_DSN",
           environment: "production",
           release: "1.0.0",
         );
@@ -78,5 +55,3 @@ class AppSentry extends AngularSentry {
     }
   }
 }
-
-```