+++ lib/angular_sentry.dart
@@ -3,129 +3,162 @@ library angular_sentry;
 import "dart:html" as html;
 import 'dart:async';
 
+import 'package:meta/meta.dart';
 import 'package:angular/angular.dart';
 import 'package:logging/logging.dart';
-import "package:sentry/sentry.dart";
+import "package:sentry/sentry_browser.dart";
+import "package:http/http.dart";
 import "package:http/browser_client.dart";
 
-const OpaqueToken sentryDsn = const OpaqueToken('sentryDSN');
+export 'package:sentry/sentry_browser.dart';
 
-@Injectable()
+typedef Event TransformEvent(Event e);
+
+/// Use to transform the event before sending to sentry
+/// add tags or extra for example
+const sentryTransformEventToken = OpaqueToken<TransformEvent>(
+  'sentry.transformEvent',
+);
+
+/// provide environment data to the sentry report
+const sentryEnvironmentToken = OpaqueToken<String>('sentry.env');
+
+/// The release version of the application.
+const sentryReleaseVersionToken = OpaqueToken<String>('sentry.release');
+
+/// Pass Logger to sentry
+/// If no logger, it will print exception to console
+const sentryLoggerToken = OpaqueToken<Logger>('sentry.logger');
+
+/// Provide sentry dsn
+/// If no dsn provided, it will log the exception without reporting it to sentry
+const sentryDsnToken = OpaqueToken<String>('sentry.dsn');
+
 class AngularSentry implements ExceptionHandler {
-  SentryClient _sentry;
-  Logger _log;
+  final Logger log;
+  final String environment;
+  final String release;
+  final String dsn;
+  final Client client;
+  final TransformEvent eventTransformer;
 
-  Logger get log => _log;
+  final _exceptionController = StreamController<Event>.broadcast();
+
+  Stream<Event> _onException;
+  SentryClientBrowser _sentry;
+
   ApplicationRef _appRef;
 
-  AngularSentry(Injector injector, @Optional() @Inject(sentryDsn) String dsn,
-      @Optional() BrowserClient client) {
-    if (dsn != null) {
-      _sentry = new SentryClient(
-          dsn: dsn,
-          compressPayload: false,
-          httpClient: client ?? new BrowserClient());
-    }
-
-    _log = new Logger("$runtimeType");
+  AngularSentry(
+    Injector injector, {
+    @Optional() this.client,
+    @Optional() @Inject(sentryDsnToken) this.dsn,
+    @Optional() @Inject(sentryLoggerToken) this.log,
+    @Optional() @Inject(sentryEnvironmentToken) this.environment,
+    @Optional() @Inject(sentryReleaseVersionToken) this.release,
+    @Optional() @Inject(sentryTransformEventToken) this.eventTransformer,
+  }) {
     // prevent DI circular dependency
-    new Future<Null>.delayed(Duration.ZERO, () {
+    new Future<Null>.delayed(Duration.zero, () {
       _appRef = injector.get(ApplicationRef) as ApplicationRef;
     });
+
+    _onException = _exceptionController.stream
+        .map(
+          transformEvent,
+        )
+        .where(
+          (event) => event != null,
+        )..listen(
+            _sendEvent,
+            onError: logError,
+          );
+
+    _initSentry();
   }
 
-  void onCatch(dynamic exception, [dynamic stackTrace, String reason]) {
+  void _initSentry() {
+    if (dsn == null) return;
+
     try {
-      _send(exception, stackTrace, reason);
+      _sentry = SentryClientBrowser(
+          dsn: dsn,
+          httpClient: client ?? BrowserClient(),
+          environmentAttributes: Event(
+            environment: environment,
+            release: release,
+          ));
     } catch (e, s) {
       logError(e, s);
-      // do nothing;
+    }
+  }
+
+  void _sendEvent(Event e) {
+    try {
+      _sentry?.capture(event: e);
+    } catch (e, s) {
+      logError(e, s);
+    }
+  }
+
+  /// onException stream after [transformEvent] call
+  Stream<Event> get onException => _onException;
+
+  /// can be override to transform sentry report
+  /// adding tags or extra for example
+  @protected
+  @mustCallSuper
+  Event transformEvent(Event e) {
+    try {
+      if (eventTransformer == null) return e;
+
+      return eventTransformer(e);
+    } catch (e, s) {
+      logError(e, s);
+      return e;
     }
   }
 
   /// Log the catched error using Logging
-  /// Called before onCatch
-  void logError(dynamic exception, [dynamic stackTrace, String reason]) {
-    _log.severe(() => ExceptionHandler.exceptionToString(
-          exception,
-          stackTrace,
-          reason,
-        ));
+  /// if no logger provided, print into console with window.console.error
+  void logError(exception, [stackTrace, String reason]) {
+    if (log != null) {
+      log.severe(reason, exception, stackTrace);
+    } else {
+      logErrorWindow(exception, stackTrace, reason);
+    }
   }
 
+  /// log error using window.console.error
+  void logErrorWindow(exception, [stackTrace, String reason]) {
+    if (reason != null) html.window.console.error(reason.toString());
+
+    html.window.console.error(exception.toString());
+
+    if (stackTrace != null) html.window.console.error(stackTrace.toString());
+  }
+
+  @protected
+  @mustCallSuper
+  void capture(dynamic exception, [dynamic stackTrace, String reason]) =>
+      _exceptionController.add(Event(
+        exception: exception,
+        stackTrace: stackTrace,
+        message: reason,
+      ));
+
   @override
+  @protected
   void call(dynamic exception, [dynamic stackTrace, String reason]) {
     logError(exception, stackTrace, reason);
-    onCatch(exception, stackTrace, reason);
+    capture(exception, stackTrace, reason);
+
+    // not sure about this
+    // the application state might not be clean
     _appRef?.tick();
   }
 
-  /// provide environment data to the sentry report
-  String get environment => null;
-
-  Map<String, String> get tags => {
-        "userAgent": html.window.navigator.userAgent,
-        "platform": html.window.navigator.platform
-      };
-
-  /// The release version of the application.
-  String get release => null;
-
-  /// provide extra data to the sentry report
-  Map<String, String> get extra => null;
-
-  void _send(dynamic exception, [dynamic stackTrace, String reason]) {
-    final event = new Event(
-        exception: exception,
-        stackTrace: stackTrace is List ? stackTrace.first : stackTrace,
-        release: release,
-        environment: environment,
-        extra: extra,
-        tags: tags,
-        message: reason);
-    _sentry?.capture(event: event);
+  void dispose() {
+    _exceptionController.close();
   }
 }
-
-//Trace parseStackTrace(dynamic stackTrace) {
-//  if (stackTrace is StackTrace) {
-//    return new Trace.parse(stackTrace.toString());
-//  } else if (stackTrace is String) {
-//    return new Trace.parse(stackTrace);
-//  } else if (stackTrace is Iterable<String> && stackTrace.length == 1) {
-//    return new Trace.parse(stackTrace.first);
-//  } else if (stackTrace is Iterable) {
-//    final trace = stackTrace.map(parseFrame).where((f) => f != null).toList();
-//    return new Trace(trace);
-//  }
-//  return new Trace.current();
-//}
-//
-//Frame parseFrame(f) {
-//  if (f is Frame) {
-//    return f;
-//  }
-//  Frame parsed;
-//  if (f is String) {
-//    parsed = new Frame.parseV8(f);
-//    if (parsed is UnparsedFrame) {
-//      parsed = new Frame.parseFirefox(f);
-//      if (parsed is UnparsedFrame) {
-//        parsed = new Frame.parseSafari(f);
-//        if (parsed is UnparsedFrame) {
-//          try {
-//            parsed = new Frame.parseFriendly(f);
-//          } catch (_) {
-//            parsed = null;
-//          }
-//        }
-//      }
-//    }
-//  }
-//
-//  if (parsed is UnparsedFrame) {
-//    parsed = null;
-//  }
-//
-//  return parsed;
-//}